+++ ui/src/utils/styles.ts
@@ -1,0 +1,10 @@
+export const SharedStyles = {
+  button: {
+    default: "rounded-md mt-7 bg-green-500 hover:bg-green-700 text-center py-2 px-4 mt-2 w-full appearance-none text-white",
+    disabled: "bg-gray-300 hover:bg-gray-300 cursor-not-allowed"
+  },
+  input: {
+    default: "border rounded-md border-gray-100 bg-white w-full py-2 px-4",
+    invalid: "border-red-500 text-red-500"
+  }
+}
